revert changes and add clap as separate features
