+++ .github/workflows/ci.yml
@@ -18,9 +18,7 @@ jobs:
     - uses: actions/checkout@v3
     - name: Curl
       run: curl -LS https://raw.githubusercontent.com/wiseaidev/ipcap/main/data/geo_ip_city.dat --create-dirs -o ~/ipcap/geo_ip_city.dat
-    - name: Build cli
-      working-directory: ./ipcap-cli
+    - name: Build
       run: cargo build --verbose
-    - name: Run tests on library
-      working-directory: ./ipcap
+    - name: Run tests
       run: cargo test --verbose
