+++ .gitignore
@@ -1,11 +1,11 @@
 # Generated by Cargo
 # will have compiled files and executables
-*/debug/
-*/target/
+debug/
+target/
 
 # Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
 # More information here https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html
-*/Cargo.lock
+Cargo.lock
 
 # These are backup files generated by rustfmt
 **/*.rs.bk
