--- ipcap-cli/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "ipcap-cli"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-clap = { version = "4.4.12", features = ["derive"] }
-ipcap = { path = "../ipcap" }
